+++ 02-main_camp/Day35/src/app/_components/myapis/myapis-write/types.ts
@@ -1,0 +1,10 @@
+interface IPlanInputType {
+  id: number;
+  title: string;
+  startDate: string;
+  endDate: string;
+  companions: string[];
+  departureLocation: string;
+  destination: string;
+  timeline: string[];
+}
