+++ 02-main_camp/Day35/src/app/myapis/[myapisid]/page.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+export default function MyPlanDetailPage() {
+  return (
+    <div>
+      MyPlanDetailPage
+      <div>
+        <div>일정명</div>
+        <div>날짜</div>
+        <div>출발지</div>
+        <div>목적지</div>
+        <div>동행자</div>
+        <div>타임라인</div>
+      </div>
+    </div>
+  );
+}
